+++ tests/server/blueprints/api_v1/test_views_delete.py
@@ -1,13 +1,14 @@
-from cgbeacon2.resources import test_snv_vcf_path
 import json
+
+from cgbeacon2.resources import test_snv_vcf_path
 
 
 def test_delete_no_auth_token(mock_app, api_req_headers):
     """Test receiving a variant delete request with no X-Auth-Token header key"""
     api_req_headers.pop("X-Auth-Token")
-    # When a POST delete request is missing the auth token
+    # When a delete request is missing the auth token
     data = dict(samples=["sample1", "sample2"])
-    response = mock_app.test_client().post("/apiv1.0/delete", json=data, headers=api_req_headers)
+    response = mock_app.test_client().delete("/apiv1.0/delete", json=data, headers=api_req_headers)
     # then it should return not authorized
     assert response.status_code == 403
     # With a proper error message
@@ -21,9 +22,9 @@ def test_delete_wrong_auth_token(mock_app, database, api_req_headers):
     # Given that no authorized API user is present in the database
     assert database["user"].find_one() is None
 
-    # When a POST delete request with token is sent
+    # When a delete request with token is sent
     data = dict(samples=["sample1", "sample2"])
-    response = mock_app.test_client().post("/apiv1.0/delete", json=data, headers=api_req_headers)
+    response = mock_app.test_client().delete("/apiv1.0/delete", json=data, headers=api_req_headers)
     # then it should return not authorized
     assert response.status_code == 403
     # With a proper error message
@@ -37,9 +38,9 @@ def test_delete_no_dataset(mock_app, database, api_user, api_req_headers):
     # GIVEN an authorized API user
     database["user"].insert_one(api_user)
 
-    # When a POST delete request is missing dataset id param:
+    # When a delete request is missing dataset id param:
     data = dict(samples=["sample1", "sample2"])
-    response = mock_app.test_client().post("/apiv1.0/delete", json=data, headers=api_req_headers)
+    response = mock_app.test_client().delete("/apiv1.0/delete", json=data, headers=api_req_headers)
     # Then it should return error 422 (Unprocessable Entity)
     assert response.status_code == 422
     # With a proper error message
@@ -53,9 +54,9 @@ def test_delete_invalid_dataset(mock_app, database, api_user, api_req_headers):
     # GIVEN an authorized API user
     database["user"].insert_one(api_user)
 
-    # When a POST delete request specifies a dataset not found in the database
+    # When a delete request specifies a dataset not found in the database
     data = dict(dataset_id="FOO", samples=["sample1", "sample2"])
-    response = mock_app.test_client().post("/apiv1.0/delete", json=data, headers=api_req_headers)
+    response = mock_app.test_client().delete("/apiv1.0/delete", json=data, headers=api_req_headers)
     # Then it should return error 422 (Unprocessable Entity)
     assert response.status_code == 422
     # With a proper error message
@@ -73,9 +74,9 @@ def test_delete_invalid_sample_format(
     # GIVEN a database containing a public dataset
     database["dataset"].insert_one(public_dataset)
 
-    # When a POST delete request contains an invalid samples parameter
+    # When a delete request contains an invalid samples parameter
     data = dict(dataset_id=public_dataset["_id"], samples="a_string")
-    response = mock_app.test_client().post("/apiv1.0/delete", json=data, headers=api_req_headers)
+    response = mock_app.test_client().delete("/apiv1.0/delete", json=data, headers=api_req_headers)
     # Then it should return error 422 (Unprocessable Entity)
     assert response.status_code == 422
     # With a proper error message
@@ -91,9 +92,9 @@ def test_delete_samples_not_found(mock_app, public_dataset, database, api_user, 
     # GIVEN a database containing a public dataset
     database["dataset"].insert_one(public_dataset)
 
-    # When a POST delete request contains a list of samples that is not found in the dataset
+    # When a delete request contains a list of samples that is not found in the dataset
     data = dict(dataset_id=public_dataset["_id"], samples=["FOO", "BAR"])
-    response = mock_app.test_client().post("/apiv1.0/delete", json=data, headers=api_req_headers)
+    response = mock_app.test_client().delete("/apiv1.0/delete", json=data, headers=api_req_headers)
 
     # Then it should return error 422 (Unprocessable Entity)
     assert response.status_code == 422
@@ -132,7 +133,7 @@ def test_delete_variants_wrong_sample(
 
     # WHEN the delete variants API is used to remove variants for a sample that is not among dataset samples
     data = {"dataset_id": public_dataset["_id"], "samples": ["WRONG_SAMPLE"]}
-    response = mock_app.test_client().post("/apiv1.0/delete", json=data, headers=api_req_headers)
+    response = mock_app.test_client().delete("/apiv1.0/delete", json=data, headers=api_req_headers)
 
     # It should return an error
     assert response.status_code == 422
@@ -170,7 +171,7 @@ def test_delete_variants_api(
 
     # When the delete variants API is used to remove variants for one sample
     data = {"dataset_id": public_dataset["_id"], "samples": ["ADM1059A2"]}
-    response = mock_app.test_client().post("/apiv1.0/delete", json=data, headers=api_req_headers)
+    response = mock_app.test_client().delete("/apiv1.0/delete", json=data, headers=api_req_headers)
     # Then the response should return success
     assert response.status_code == 200
     resp_data = json.loads(response.data)